pyproject.toml
[project]
name = "dataforge"
version = "0.1.0"
description = "DataForge MVP platform skeleton"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pytest>=8.4.2",
    "pyyaml>=6.0.3",
]

[project.scripts]
dataforge = "dataforge.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.uv]
package = true
